move MDX utils out of codemods
